Move style of menu buttons to separate file
